[WIP] Clean up code and add new facilities
This commit moves `get` into a new environment object. The idea behind
the `get` abstraction is to allow the same validators to be used with
Ember objects and in environments without Ember.get without having to
write special validators for each environment.

Additionally, this commit makes it possible for an environment to
register validators by name. The test environment uses a simple
registry (register()), but an Ember environment would probably use
container to look up the validators.

Finally, this commit adds an initial implementation of a `length`
validator in the tests, which required adding a new `getSubProperty`
method on `SingleFieldValidator`. As above, `getSubProperty` allows the
same validators to be used both in Ember (with Ember.get) and in other
environments.

We currently intend for validators like the `length` validator to do
nothing if the field isn't an object with a `length` property. If you
also want to validate that the field is present, you can pair it with
the presence validator. This ensures that a given field doesn't report
multiple errors unnecessarily.

To make this nicer to use, we could write helpers like:

```
function length({ min, max }) {
  return [
    validates('presence'),
    validates('arraylike'),
    validates('length', { min, max })
  ]
}

dsl({
  email: length({ min: 1 })
});
```

This makes it possible to express `{ allowBlank: true }` and `{
allowNotArray: true }` easily without needing to add those kinds of
options everywhere.