Add a more ergonomic example to tests
This test illustrates that it's possible to easily create validators
that combine other validators together.

This is important because at the primitive level, you have to be fairly
verbose, narrowing down the types before you get to the meat of your
validation. This ensures that you don't get duplicate errors for a field
(email is not null, email is not a string, email is not an email), but
in practice, the validators people will use should mostly be pre-fab
ones that combine things together.

The nice thing about keeping the primitive validators simple is that if
someone wants to validate that an email is a valid email, but doesn't
care if the email is blank, all they need to do is use the string and
email validator and leave off the blankness validator.

I considered being a little less principled about this, but it just
meant a lot more case-by-case thinking on a per-validator basis.